requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
psycopg2-binary==2.9.6
python-dotenv==1.0.0
cryptography==41.0.3
PyJWT==2.8.0
passlib[bcrypt]==1.7.5
pydantic==1.10.11
pytest==7.4.0
httpx==0.24.1
pytest-cov==4.1.0
